requirements.txt
# CycleGAN Project Requirements
# Author: RSK World
# Website: https://rskworld.in
# Email: help@rskworld.in
# Phone: +91 93305 39277

torch>=1.9.0
torchvision>=0.10.0
numpy>=1.21.0
Pillow>=8.3.0
matplotlib>=3.4.0
scikit-image>=0.18.0
tqdm>=4.62.0
tensorboard>=2.6.0
jupyter>=1.0.0
ipywidgets>=7.6.0
scipy>=1.7.0
lpips>=0.1.4
flask>=2.0.0
flask-cors>=3.0.10
pytest>=7.0.0

